Create temporary files in binary mode on Windows
ArchMakeTmpFile previously created temporary files in the
default text mode on Windows, which caused line endings
to be translated to CRLF when writing to the file. It
now creates files in binary mode instead to maintain
consistent behavior with Linux and MacOS, which have no
such translation mechanism.

The two consumers of this in Tf are TfAtomicOfstreamWriter
and TfSafeOutputFile. TfAtomicOfstreamWriter is unaffected
by this change, since it explicitly closes then reopens
the temporary file in binary mode. TfSafeOutputFile has
been updated to also reopen the temporary file in binary
mode.

Ultimately, we may need to be able to specify whether a
temporary file is created in text or binary mode.

(Internal change: 1872279)