[UsdMaya] Change assembly edit util to only return relative edit paths
We don't handle edits with absolute paths properly, so we need to mark them as invalid
instead of parsing and returning a weird edit. The docs for the edit util have been updated
to reflect this.
(It seems that edits with absolute paths are created when there are parenting/
reparenting shenanigans. We might be able to consider those when parsing edits,
but that requires a bit more experimentation.)

(Internal change: 1872719)