[HdSt] Refactor ShaderTerminal to separate Repr Style and SurfaceShader Opinions.
The fragment shader function ShaderTerminal() has a tricky job of needing to resolve the various options of what the fragment color should be.

The SurfaceShader opinion provides the color that would be expected under normal conditions, but various other opinions exist to help with interactive rendering:
- The primvar "color".
- Display style associated with the "repr".  Which controls things like, "wireframe", "outline" and "sheer" styles.
- The primvar "overrideColor" managed by the renderpass.
- Selection color
- "Scalar Override", which is a imaging mode for data visualization.

These different opinions interact in complex ways, but ultimately it is the repr's display style that controls these interactions.

Another tricky factor, is that care must be taken on if the surfaceShader() function is called, as it can have side effects.  In particular, certain surface shaders contain "discards", which isn't desirable behavior with certain other combinations of state.

Currently, there are some bugs associated with the interactions between the scalar override and repr style, so this change lays some
foundations for fixes those issues.

The main purpose of this change is the separation of the display style opinion, from the operations of fetching other opinions and combining those opinions.

With this change the _ShadingTerminal() terminal has been removed in favor of a new terminal GetReprStyle().  This terminal returns a structure that controls hows the different opinions are combined.  Therefore the repr mixins are no longer responsible for fetching and combining these opinion, this is now the responsibility of the common ShadingTerminal() function.

Changes:
- Changes the order of the inclusion for the common and repr ("terminal") mixins.  This is because the common mixin defines the ReprStyle structure the other mixins return.
- Replaced _ShadingTerminal() terminal with  GetReprStyle()
- ShadingTerminal() is now responsible for gathering color opinions and combining them based off the repr style.
- Due to the now common Shader Terminal several of the repr mixins that didn't define integrateLights() now need to, as although the function isn't called it will be referenced, so produces link errors if not defined.  The implementation is to call integrateLightsConstant, which does a simple initialization of the return structure.

(Internal change: 1872737)