[Hdx] Fix for DrawTarget's being too restrictive about possible render tags.
Hydra currently has two mechanism where by the render tag list is filtered.  The first is via the collection, the second is via the render pass.

In the case of draw targets, the draw target render pass was providing a very narrow filter for what should be rendered (geomentry and interactive only geometry).

In most cases this what is desired, but we have an internal feature that would would prefer to use a different set of tags.

Therefore, this change removes the filter completely from the draw target render pass.  Internally, these filters were moved to rprim collections we pass to the draw target sprim.

This is  more efficient from the hydra perspective and the collection render tag filters is much earlier in the sync and execute pipes.

Changes:
- Removed DRAW_TARGET_RENDER_TAGS from HdxDrawTargetRenderPass::Execute() and we now pass the empty tag set to RenderPass::Execute().

(Internal change: 1872875)