Added a simple commandline utility for creating a ".usdz" package from a given list of files.
This initial version of the script does not attempt to localize external references in
the given files and modify the files in any way. It simply adds all files to the archive
in the specified order. Files in sub-directories are only added if the "-r" flag is specified.

For proper functioning of the output usdz file, the root USD layer must be specified first
in the list of input files.

(Internal change: 1873324)
(Internal change: 1873453)
(Internal change: 1873477)