Validate textures via resolved path in usdImagingGL
usdImagingGL now checks whether the resolved path
field in SdfAssetPath is empty to determine if a texture
exists instead of calling TfPathExists on that field. This
was redundant, as the resolved path is only populated
if the asset path resolved successfully. In addition, with
the new ArAsset API, checking TfPathExists is overly
restrictive, since it disallows textures that may not
exist as individual files on a filesystem. For example,
checking TfPathExists would disallow reading textures
packaged within a .usdz file.

(Internal change: 1873396)