Add resolver indirection to ArGetResolver
ArGetResolver now returns an instance of a private resolver
implementation that wraps around and forwards calls to
the client's resolver. This will be used in a forthcoming
change to add resolution for 'package' assets.

This additional level of indirection allows us to overlay
new functionality while minimizing impact on
custom resolvers. However, this is a breaking change
for consumers that had relied on dynamic_cast'ing
the result of ArGetResolver to their underlying resolver
subclass, or getting back their subclass when calling
Ar.GetResolver() in Python. To help with this, a new
method ArGetUnderlyingResolver has been added
that will return what ArGetResolver had previously
returned.

(Internal change: 1873779)