Add "package asset" concept to Ar
Ar defines the concept of a "package asset" as an asset
that contains other assets. To support this:

- Ar defines a "package-relative" path syntax for referring
  to assets within a package asset. An example of this is:

     "/dir1/dir2/package.usdz[subdir/file.usd]"

  Utility functions for parsing these paths have been
  added in packageUtils.h.

- Ar has a new ArPackageResolver interface for defining
  "package resolvers". These package resolvers are
  responsible for resolving and retrieving assets within
  their associated package format.

Package-relative paths may be passed to all functions
on the ArResolver object returned by ArGetResolver.
This object will detect and parse package-relative paths
and dispatch to the client's underlying resolver or
package resolver as necessary. Custom resolver
implementations should not need to be modified to
accommodate these paths.

These changes allow clients to work with package assets
and package-relative paths without explicitly having to
check whether they are dealing with a package asset.
For example, clients can pass in a path to the OpenAsset
API to open the specified asset regardless of whether
that path is a normal asset path or a package-relative
path.

(Internal change: 1873939)
(Internal change: 1873984)
(Internal change: 1874081)