[HdSt/Hdx] Fix for repr Sheer + Scalar Override
This change is part of a series of checkins for fixing interactions between scalar override and repr draw style.  In this case, the repr in question in Sheer.

The scalar override mode is meant to just effect the rgb channels, when running this mode we want to the alpha channel be controlled by the display mode.  Though it is currently undesirable for the surfaceShader to run, as it can contain discards, which are undesirable in scalar override mode.

The code to calculate the scalar override opinion has been factored out of ApplyColorOverrides() and now returns a structure containing an enable flag and the scalar override color.

ShadingTerminal() then combines the repr style and if the repr style request, applies the color overrides.

There are a couple of other subtle changes, that are not obvious from the diffs:

- The overrideColor primvar is now applied to scalar override.  Scalar override and override color don't have any known interaction, so this change was made as a simplification (to avoid having to split ApplyColorOverrides into 2 different terminals).

- The color ramp scalar override no longer modifies alpha.

Changes:
- Added interface for GetScalarOverride() to CommonTerminals
- ShadingTerminal, now calls GetScalarOverride() and if enabled it applies the alpha and color override as per reprStyle to the result.
- scalar override function moved to it's own function in CommonTerminals.
- The "mix" for the color ramp in scalar override is now rgb rather than rgba.
- GetScalarOverride() sets the enable return based on the presence of the scalarOverride primvar.

(Internal change: 1873958)