[usdMaya] preserve unsigned-ness of index in PxrUsdMayaRoundTripUtil::MarkAttributeAsArray()
Casting 'index' to an int before saving it in the custom data caused its
type name in USD to be signed. When we then tried to read it back as
unsigned, it would fail because the VtValue we get out of the VtDictionary
is holding unsigned int, not int.

(Internal change: 1874058)