[Hd] Remove bogus verify in HdResourceRegistry::GetResourceAllocation()
Recently, we changed the Hydra texture system, to provide better error and fallback handling of texture issues.  This meant that error states were carried through the system, often by the presence of nullptrs.

Therefore, the verify for texture resource being null in HdResourceRegistry::GetResourceAllocation() is no longer valid, as this is now a valid (even though indicating an error) state for texture resource to be in.

Changes:
- Removed the verify that texture resource is not null (leaving the if check).

(Internal change: 1874193)