tf: move Tf smart pointer NULL pointer dereference fatal error out of line
The amount of code generated for Tf pointers' operator-> can be an
impediment to inlining.  This change moves as much of the fatal error
code as possible into a common helper function without sacrificing
fidelity in the error message.

We generate a static call context because it allows us to trade some
space (read-only data generated during compile time) in exchange for
shaving a few more instructions off operator->.

(Internal change: 1874293)