Enhancements to usdzip commandline utility:
* Made the "inputFiles" argument optional and added an optional --asset argument,
which can be used to create an archive containing a given asset and all of
its external dependencies.

* Replaced the --quiet flag with a --verbose flag, thereby making quiet mode the
default.

* Added a -l (or --list) option to list the contents of the given usdz file to stdout or
to a specified file. If the usdz file is being created by the run, then the contents (file
list) of the created package are output in the end.

* Added a -d (or --dump) option to dump the contents of the given usdz file to stdout
or to a specified file. The dumped contents include the file offsets, compression info
and the names of files in the archive.

With the addition of the -l and -d options, the script can now be used to simply inspect
existing usdz files.

(Internal change: 1874594)
(Internal change: 1874732)