[usdMaya] add a USD stage notice listener and use it to trigger proxy shape redraws in Viewport 2.0
This change adds a simple USD stage notice listener class that can listen
for notices on a given USD stage and invoke registered callbacks in
response. The primary use for this at the moment is to ensure that USD
authoring operations that may occur outside of Maya's knowledge have a way
of notifying Maya that proxy shapes that reference that stage may be dirty
and need to be redrawn in Viewport 2.0 (where the draw override for the
proxy shape is constructed with isAlwaysDirty=false).

Currently, the only notice being listened for is StageContentsChanged,
which is the most conservative notice sent when a stage is modified. The
stage notice listener class can be extended in the future to support other
stage notices and/or to optimize notice listening to minimize unnecessary
redraws.

(Internal change: 1875035)