Fix package resolver issues with certain build types
Ar had previously been assuming that it would always be
able to find the appropriate plugin and construct an
instance of a given ArPackageResolver subclass if the
TfType for that package resolver was found. However,
this is not always the case. For example, in static
builds, a program might discover the declaration of
a package resolver's type in a plugInfo.json, but not
have linked against the library providing the
definition.

The code to discover and instantiate package resolvers
now silently skips over cases where a plugin couldn't
be found or a resolver couldn't be instantiated. This
is similar to how file format plugins are handled in Sdf.

(Internal change: 1875039)