[Hd/HdSt] Fix for Scalar Override and repr's Outine & Surface+Hull
This change is the next in a series of checkins for fixing interactions between scalar override and repr draw style.  In this case, the the two repr's being addressed are Outline & Surface+Hull.

The commonality between these two repr's is that they are multi-pass repr's.  The issue is that for for multipass repr's we typically want to apply scalar override to the first pass only.  This can't be a repr draw item control like the other changes as, for example, the hull repr draw item mixin is used as a first pass and as a second pass.

Rather than hard code, that scalar override is a first pass only thing in Hydra, it has been made a configurable attribute of ReprDesc.

The GetScalarOverride() function, that was introduced in the last check-in has now been moved to it's own mix-in implementation, Fragment.ScalarOverride, with a new implementation of that function in an alternative mix-in Fragment.NoScalarOverride, which disables the scalar override function.

This means all the shader keys needed updating to include one of the variations  of the mixin.  For points prim-type, we always disable scalar override.  For curves, the original functionality is retained (i.e. it's enabled if the scalarOverride primvar is present).

For meshes, the shader key has a new boolean control that selects between the two variations.  This value is driven by a similar new control in HdMeshReprDesc.

Changes:
- Added enableScalarOverride to HdMeshReprDesc.
- Mesh.cpp now passes enableScalarOverride from the repr desc to the shader key.
- Increased the fragment shader token counts in all the shader keys to account for new mixin.
- Added the Fragment.ScalarOverride mixin to the HdSt_BasisCurvesShaderKey.
- Added the Fragment.NoScalarOverride mixin to the HdSt_PointsShaderKey.
- HdSt_MeshShaderKey now adds either the Fragment.ScalarOverride or Fragment.NoScalarOverride mixins based off the enableScalarOverride flag passed in.
- In terminals.glslfx, move the implementation of GetScalarOverride() to new mixin Fragment.ScalarOverride and added a mixin Fragment.NoScalarOverride that implements GetScalarOverride() to return a ScalarOverride control structure, with scalar override disabled.

(Internal change: 1875142)