[usdMaya] Support for reading typeless defs and prims with no prim reader
When re-exported, prims with unknown type name are treated as if they were typeless defs
(the type information is not preserved).

The UsdMayaTransformWriter is updated to explicitly allow Maya transforms to be
exported to non-UsdGeomXformable prims (to accommodate the use of Maya
transforms as typeless defs and other unknown prim types).

(Internal change: 1875469)