[Glf/Hd/HdSt] Add optimization to defer load textures.
When switching to hydra imaging, a couple of tests regressed loading performance.  The reason for this, is these test scenes contained textures that wasn't actually used in scene (they were wired to shaders, but the shaders were disabled).  The time to load these textures accounted for the additional time in the performance tests.

In order to fix this, this change modifies the way textures are loaded.  Previously, textures were loaded whenever the memory request was changed (which is done when the HdStTextureResource is created).  Instead, textures are now loaded on demand, when one of the getter API's that depend on the loaded file are called.  As a result a lot of the getters lost their const-ness.

A special case of this is GetTextureInfo() mechanism, which is used both to query the current state for profiling and to obtain actual values from the file from processing.  Therefore, this method has a forceLoad flag added to it.

As SetMemoryRequested() no longer triggers the load, the internal _OnSetMemoryRequested() method has replaced with a dirty notification function,  _OnMemoryRequestedDirty.  This method typically just clears the loaded flag on the object.

The deferred loaded code is typically handled by GlfBaseTexture.  In classes that are derived from this class, their OnSetMemoryRequested() method has been replaced by the _ReadTexture() function.

So the process goes:
Create the texture -> Loaded flag set to false.
SetMemoryRequested -> _OnMemoryRequestedDirty() -> Loaded flag set to false.
Texture Getter (such as GlTextureName, width, height, format) -> If Loaded flag is false -> _ReadTexture() -> Loaded flag is set.

There was two fallouts from this change:
- Changes to decrease the memory requests were not reflected until the next time the shader was dirty.
- Texture Resource Constructor doesn't have the wrap modes available to create the samplers.

The first problem is caused by an order of operations issue.  We would sync the Bprim's which would add the new memory request to the texture, then update the Sprim's, which would cause the texture to evaluate it's state.  At this point the texture had both the old and new memory request, so would get loaded with the largest of the two.  We would then garbage collect the old texture in commit resources, which would remove the memory request.

The fix to this problem, is to make Bprim garbage collection a step in its own right, that happens between Bprim and Sprim sync.

The second issue is fixed by also deferring sampler creation.

Changes:
- GlfTexture interfaces for GetBindings() and GetTextureInfo() made non-const.
- GetTextureInfo now takes a boolean flag to force loading.
- _OnSetMemoryRequested() replaced with _OnMemoryRequestedDirty().
- GlfBaseTexture's getters, GetGlTextureName(), GetWidth(), GetHeight(), GetFormat() no longer const or inline.  They now cause the texture to be loaded if needed.
- GlfBaseTexture implements OnMemoryRequestedDirty() to clear the new loaded flag.
- GlfBaseTexture has new api to read textures (_ReadTexture(), pure virtual) and to set the loaded flag (_SetLoaded)
- GlfTextureRegistry::GetTextureInfos() calls to GetTextureInfo() don't cause the textures to be loaded.
- GlfPtexTexture doesn't use GlfBaseTexture, so has a similar set of changes.
- The change tracker now has a separate garbage collection bit for Bprims, that the global one.
- In render index sync, we now garbage collect Bprims after their sync step.
- In resource registry, separated out Bprim garbage collect from global garbage collect.
- Texture Sync now causes Bprim specific garbage collection to run instead of the global one.
- HdStSimpleTextureResource constructor no longer creates the sampler, but instead stores in the class the passed in wrap modes.  The getter for sampler id, now creates the sampler.

(Internal change: 1875916)
(Internal change: 1876072)
(Internal change: 1876112)