[usdMaya] Fix some naming issues in readJob, writeJob, writeJobContext
UsdMayaReadJob and UsdMayaWriteJob are private classes, so we can do any renaming
with impunity.

UsdMayaWriteJobContext is public, but wasn't until somewhat recently. So the thought is
that we should rename and fix the access of class members before people actually start
depending on the API. (The affected API would probably be accessed through methods
on UsdMayaPrimWriter anyways. Note that we didn't have to change any of the prim
writers in pxrUsdTranslators.)

(Internal change: 1876045)