[HdSt/UsdImaging/UsdImagingGL] Don't load custom shaders when "Hardware Shading" is disabled.
While the "Enable Hardware Shading" option in UsdView, did disable the application of custom shaders, the shaders where still loaded by usdview and the texture dependencies were loaded and downloaded the GPU (with the stream backend).

The first part of this change, is to make Hydra backends responsible for it's own fallbacks. Similar to the previous change to texture fallback handling, in this case, the HdSt material code, will fallback to the fallback surface shader, if their is an issue with user provided one. While UsdImaging never implemented the fallback mechaism, other front-ends were querying render delegate specific code to obtain the correct fallback. Now, if both the surface and displacement shader sources are empty strings, then the fallback is used.

The second part of this change, is to wire through the "Enable Hardware Shading" ui option to the UsdImaging Scene Delegate. When the toggle is changed, we need to invalidate all the delegate material bprims, so that Hydra picks the change. This is done via the adapters, in a similar mechanism to updating root transforms.

Then, we then scene delegate receives shader source requests, if hardware shading is disabled, it returns an empty string, triggering the backends fallback behavior.

Changes:
- HdStMaterial not contains a single static _fallbackSurfaceShader, which is initialized on first use. If both the surface shader and displacement shader sources are empty, the fallback surface shader is used instead.
- UsdImagingDelegate has new api for calling SetHardwareShadingEnabled(), to change the enable state.
- In the new SetHardwareShadingEnabled() api, if the enable state changes, visit all the adapters notifying them that the shader source is dirty.
- In GetSurfaceShaderSource(), GetDisplacementShaderSource() and GetMaterialParams(). If the hardware shading is dirty return an empty object.
- In UsdImagingPrimAdapter added new API for MarkMaterialDirty. The default implementation is a noop.
- In Material, adapters, MarkMaterialDirty, sets the DirtyResource dirty bit on the sprim.
- HydraMaterial and DrawMode adapters, MarkMaterialDirty, sets the DirtySurfaceShader and DirtyParams dirty bits on the sprim.
- UsdImagingGLHdEngine::Render() now forwards the hardware shading enable from render params to the delegate.

(Internal change: 1876287)