[usdviewq] Fix triggered slot for certain QActions
Selecting an item from the "Renderer" or "Camera" menus
in usdview would cause an error because the triggered()
signal for the associated QAction was connected to a
function that did not accept the 'checked' bool
parameter. This seems to have worked in previous
versions of PySide2 but does not work with the
latest pip-installable release.

(Internal change: 1876369)