Replaced implicit conversion operators that allow auto-conversions between ConnectableAPI and Shader or NodeGraph and vice versa with non-explicit, converting constructors with python wrappings.
Removed calls to boost:python::implicitly_convertible that define implicit conversions from a
ConnectableAPI object to a Shader or NodeGraph object. These are conceptually a downcasting
operation and aren't being used anywhere. This avoids the issue reported in bug 162470, where
boost::python was picking the wrong converter for a NodeGraph (or Material) object when
attempting to convert it to a ConnectableAPI, resultingin infinite recursion.

(Internal change: 1876748)