[px_vp20] remap the kShadowBias value back into the light's bias attribute value so it can be used by Hydra
Maya's default value for the "Bias" attribute on lights is 0.001, but when
querying the draw context for lighting data, that value gets reported by
the MLightParameterInformation as 0.0022. When the attribute is set to
-1.0, 0.0, or 1.0, it is reported back to us as -0.198, 0.002, and 0.202,
respectively, so from that we derive the function to remap the reported
value back to the attribute value.

(Internal change: 1877057)