[build_usd] fix for building usd from a symlinked directory
Currently (on MacOS anyway, haven't tested on linux), invoking
build_usd.py from inside a symlinked directory results in usd installing
in the incorrect location - the dependency install dir (if you've
specified a separate one, with --inst).

This fixes this by explicitly storing which dependency we're building on
the context, instead of trying to implicitly figure it out by doing a
text comparison of the current directory.