[usdMaya/pxrUsdMayaGL] ignore PIXMAYA_ENABLE_BOUNDING_BOX_MODE in proxy shape draw override
This change removes the public UsdMayaIsBoundingBoxModeEnabled() function
in the USD proxy shape and leaves the PIXMAYA_ENABLE_BOUNDING_BOX_MODE
environment variable only used internally to the proxy shape in its
isBounded() method. The env var is basically never enabled, so proxy shapes
were effectively always appearing unbounded to Viewport 2.0.

The proxy shape advertises itself as unbounded to gain a performance
benefit, as we're able to get Maya to bypass its own costly CPU-based
frustum culling in favor of Hydra's higher performance implementation.

This became problematic though for features in Viewport 2.0 such as
automatic computation of the width focus for directional lights, since the
draw override cannot get a bounding box for the shape. By claiming to be
unbounded, proxy shapes were not having their bounding boxes accounted for
in computing the width of the light, resulting in improperly fit depth maps
for directional lights.

Since the intent of PIXMAYA_ENABLE_BOUNDING_BOX_MODE is to get Maya to
bypass its own drawing of bounding boxes for proxy shapes, we can achieve
that for Viewport 2.0 instead by overriding the
disableInternalBoundingBoxDraw() method of MPxDrawOverride.

(Internal change: 1879109)