[refactor] make the repr tokens a concrete type in preparation for allowing sets of tokens as reprs.
This changes HdRepr selection from using raw TfTokens to a struct
called HdReprSelector. We are the refactor to allow the selector to
pick multiple representations. The API aims to reduce the reliance
on specific TfTokens for proper repr usage as they are application
specific should not be referenced in individual backends.

(Internal change: 1879110)