[usdMaya] Add ARKit compatibility mode for USD export
Introduces general flag "compatibility" for the export command and translator
for instructing the export process to change behavior according to the current
compatibility profile. The notion of "compatibility profile" is intended to encapsulate
opaque changes to the export process that we currently don't intend to expose
to the user as separate options.

Compatibility profile "standard" is the preferred or default profile that provides the
"canonical" export.

Compatiiblity profile "appleArKit" uses the ARKit-specific packaging function in
UsdUtils and uses the usdz extension as fallback if no extension is specified.

Other compatibility profiles can be added in the future.

(Internal change: 1879159)