Handle outputs outside of a nodegraph correctly.
We used to put nodes and outputs outside of a nodegraph into
an "ad hoc" nodegraph.  This was easy but in USD shaders in a
nodegraph in a material have different semantics than shaders
that are direct children of the material.  So now we leave them
outside a nodegraph and we have to look up their outputs in a
different way.

This also fixes finding shader ids for standard node types when
the standard library isn't explicitly included.  We simply load
the standard library and look for shaders there when they're not
found in the current document.

(Internal change: 1879266)