[UsdImaging] Optimization for Usd Imaging Delegate Path Translation
The UsdImaging Delegate has 3 path namespaces: "usd path", "cache path" and "index path".

The usd path represents a path in the usd scene, where as cache path and index path represents the path of the imaging prims.  Cache path is being a path local in scope to the delegate and index path being scene global scope (used in the hydra render index).

The translation between cache path and index path involves adding or removing the delegate path as a prefix to the path.  To do that, the code invokes the SdfPath::ReplacePrefix() function, which is a well known performance bottleneck.

Previous versions of the UsdImaging delegate had a cache of these translations, which was populated on demand.  However, as the translation code is executed from multi-threads, it require locks to be added and those locks cost more than the benefit of the cache.  So the cache was removed, in the hopes that one day, the need for the translation was removed entirely.

The SdfPath::ReplacePrefix() function has returned to being a bottleneck and the removal of the translation code is still a large project.  So this change re-introduces the cache.

However, this time we take advantage of the hooks we placed to add and remove the priminfo entry to populate and depopulate the translation cache.  Making the multi-threaded accesses read-only, therefore not requiring locking on the cache.

Changes:
- UsdImagingDelegate now contains a path to path map for both cache path to index path and index path to cache path translations
- In UsdImagingDelegate GetPathForIndex() and GetPathForUsd() now first attempt to look up the paths in the translation caches, if not found falling back to SdfPath::ReplacePrefix().
- UsdImagingIndexProxy::AddPrimInfo() now add path translation entries to the caches.
- UsdImagingIndexProxy::_ProcessRemovals() removes path translation entries from the caches.

(Internal change: 1879508)