[Hd] Optimization to not run garbage collection if a texture hasn't change.
This is an optimization for a case where in one of Hydra's scene delegates, the texture resource dirty bit is being set, but the resource hasn't really changed (the Id returned is identical).

If this happens, we were running the garbage collector, but this unneeded.  While a previous change made the garbage collection step cheaper, it still an unnecessary operation.

Changes:
- HdTexture::Sync keeps a flag to see if garbage collection is needed (isNewTexture).
- If this isn't the first time we've seen the texture, it maybe the same, so check if the resource pointer is different.  If it is the same, it's not a new texture.
- Only run the garbage collector for new textures.

(Internal change: 1879554)