[pxrUsdMayaGL] set "constant" render setup task params at task creation and minimize param dirtying
A number of parameters set on the HdxRenderTaskParams for the render setup
task were being set to the same value in every call to
PxrMayaHdSceneDelegate::GetRenderTasks(), so instead we set those once when
the task is inserted into the render index.

We also only mark the render setup tasks' params dirty if one of the
"varying" parameters was modified.

(Internal change: 1879592)