[pxrUsdMayaGL] separate old "queue info" debug code into drawing and selection codes
This change replaces the PXRUSDMAYAGL_QUEUE_INFO debug code with
PXRUSDMAYAGL_BATCHED_DRAWING and PXRUSDMAYAGL_BATCHED_SELECTION debug
codes. It also cleans up the debug output a bit and includes additional
info for the frame stamp and pass context when working with Viewport 2.0.

Finally, the _drawnMayaRenderPasses member variable was removed, as it is
no longer necessary now that we use the frameStamp from the MDrawContext
for Viewport 2.0. The frame stamp ensures that we only draw a particular
Maya render pass once.

(Internal change: 1879596)