[UsdImaging/UsdImagingGL/UsdSkelImaging] Fix for missing primvars when enabling hardware shading in usdview.
A previous changes added support for hydra/usdimaging to not load shaders and textures as the hardware shading (i.e. custom shading) option was toggled.  However, if usdview was started with hardware shading disabled and then later enabled, the shaders were still not right and it looked like textures were missing.

It turned out the problem was that Primvars were missing.  These are managed by the rprim, so when the material is marked as dirty, the rprims must also be marked dirty to inform them to re-evaluate the primvar list.

Changes:
- All Rprim-based adapters now also implement MarkMaterialDirty, to set DirtyMaterialId.
- Added some missing USDIMAGING_API from UsdImagingGprimAdapter.

(Internal change: 1879674)