[UsdGeom] Speed up extent computation.
Parallelize extent computation using WorkParallelReduce.

This proved to provide non trivial speedups in one of our
file format plugins which needed to compute extents for every
mesh it was creating (in the realm of 30% on a 150MB asset with
meshes accounting for 30% of its total number of prims).

For grain size, there was discussion regarding the value chosen here:

https://github.com/PixarAnimationStudios/USD/pull/588