Improved documentation for the -l and -d options to usdzip. If an existing usdz file is specified as an argument to the -l or -d option, it now acts as the input 'usdzFile' argument (which is now optional according to the argument parser). In this case, the listed / dumped contents are output to stdout.
Added a new TF_DEBUG symbol USDUTILS_CREATE_USDZ_PACKAGE which when enabled
outputs messages about the asset being packaged and the files being added to
the package.

When the --verbose option is passed into usdzip in pacakge creation mode, the
debug symbol is enabled so that useful information regarding the packaging of
an asset is output to stdout.

(Internal change: 1881085)