Modify the testSdrRegistry test to handle monolithic builds.
In a monolithic build we'll find all linked-in ndr discovery and
parser plugins but the test expects to find only the ones in a
specific plugin search path.  This change handles that in two
ways.

First, we simply accept that we might find more plugins than
originally expected.  We still require that we find the ones
that were originally expected.

Second, we disable normal discovery and explicitly tell the
registry which discovery plugins we want.  Those would be the
plugins originally expected.  Since, AFAIK, there's no way in
Python to create a discovery plugin from a type we can't use
the SetExtraDiscoveryPlugins() as is.  This change adds an
overload that takes a vector of TfType where each is the type
of a discovery plugin.  In C++ this creates the plugins using
the factory in each TfType then passed them to the original
SetExtraDiscoveryPlugins().  The Python wrapping now allows
callers to pass a vector with both plugins and types.

(Internal change: 1882533)