[Hd/HdSt/usdHydra/usdImagingGL] Update wrap mode fallback behavior.
The original implementation for texture wrap modes in Usd and Hydra, was based on our animation tool.

In it, a Texture prim can define a wrap mode and if it doesn't have an opinion, we defer to the meta data in the texture file itself.  If that texture file doesn't define a wrap mode, then it uses Repeat, as this was basically OpenGL's default.

The Renderman spec slightly differed from this, saying Black was the fallback when neither the prim nor texture file provide one.

In a recent decision, UsdShade was going to align with Renderman and have Black as a fallback, but didn't provide a way that let the texture file to have an opinion.

Therefore, it is desirable to introduce the missing mode of using the texture file metadata to Usd and standardize on Black being the fallback value everywhere.

However, there is some concern about existing assets that have been created and may depend on the current behavior.  Therefore, it was decided that we continue to use the legacy (repeat) fallback behavior for HwUvTexture_1 shader nodes and use the new fallback (black) for UsdUVTexture shader nodes.

Changes:
- Changed HdWrapUseMetaDict to HdWrapUseMetadata, so that the Hydra name matches the new Usd token name.  This enum value also indicates following the new desired fallback behavior (i.e.  black).
- Added HdWrapLegacy wrap mode which matches the current HdWrapUseMetaDict fallback behaviour (i.e. repeat).
- Added "useMetadata" UsdShade token to UsdHydra.
- Filtered _GetWrapS and _GetWrapT in UsdImagingGL's textureUtils.cpp into a a single function.
- In the new wrap function, if the usdPrim doesn't specifies a wrap mode or its set to useMetadata, we use one of the Hd metadata values depending on the prim's id attribute.

(Internal change: 1882744)