[Hd/HdSt] Fix for incorrect draw targets being used when switching stages and back.
In an internal tool, we had a bug where the same model was loaded into two different stages (two render indexes in the hydra sense).  The stages use a draw target and when the user switch stage and then back again, the draw target looked incorrect.

When Hydra Stream creates a draw target, it asks the scene delegate for a Unique texture id for that draw target.  It then uses that Id to add the draw target's texture to the resource registry's texture registry.  As a draw target has no content to create a unique id on, the tool's scene delegate creates the id from hashing the attachment path (a combination of the draw target prim id and attachment name).

The root cause of the bug, is the same model is used in both stages, so the path and thus id's are identical for both stages.  As the resource registry is shared between render indexes, it caused a collision in the texture registry, so the first stage set up it's draw target texture resource, the second stage updated the entry to its draw target texture resource, so when returning to the original stage, the draw target was writing to the original texture resource, but Rprim in the main render pass that used the results was picking up the draw target from the second stage.

Changes:
- Changed _TextureResourceRegistry key from HdTextureResource::ID to HdResourceRegistry::TextureKey to indicate this value is no longer the id returned by the scene delegate.
- HdResourceRegistry texture registry accessor functions RegisterTextureResource() and FindTextureResource(), now take the new TextureKey.
- A new method, GetTextureKey() has been added to the render index, to convert the key from
  render index local namespace to global.  (by salting the hash of the id with the hash of the render index this pointer)
- HdTexture, HdStDrawTarget and HdStMaterial have been updated for the new api.

(Internal change: 1882800)
(Internal change: 1883308)